Migrate Rust integration tests to Ruby CLI and Rust unit tests
Rust integration tests tested the CLI pipeline through internal APIs,
making them fragile to internal refactoring. Move error detection and
location tests to Ruby CLI tests for true end-to-end validation, and
keep type inference tests as Rust unit tests where they belong.
