+++ rust/src/analyzer/mod.rs
@@ -7,7 +7,4 @@ mod literals;
 mod parameters;
 mod variables;
 
-#[cfg(test)]
-mod tests;
-
 pub use install::AstInstaller;
