--- rust/src/analyzer/tests/mod.rs
@@ -1,1 +1,0 @@
-mod integration_test;
